fix: Duplicate `pub mod` definitions in machine-a-tron
We were doing the same `pub mod` in main.rs and lib.rs so that all
machine-a-tron could be compiled either as a library or as a binary, but
it quickly becomes tiring having to duplicate the same definitions in
both files.  Plus it prevents us from putting any utility code in lib.rs
(because main.rs can't see it).

Make lib.rs the source of truth for everything not in main.rs, and only
make things public that are actually used outside the lib (main plus the
integration tests.)