fix: Drop WaitForSiteExplorer state from machine-a-tron
This was a workaround for a bug where carbide would generate incorrect
predicted host info when a host would run discovery before site-explorer
found it. I justified avoiding the bug in MAT because (a) it was already
fixed in an open branch (Abhishek's multidpu state machine MR) and (b)
booting machines before site explorer is something we could
realistically avoid in real life by simply not booting machines until
site explorer runs (meaning it's not so awful to do the same in the
test.)

Now that the bug is fixed (abhi's MR is merged), we can drop the
WaitForSiteExplorer state and boot machines right away.