fix: do not update DPUs in a multi-dpu host if multi-dpu support is disabled
Until the reprovisioning work is complete, auto-updater does not support multiple dpu hosts.  In order to use auto-updater for single DPU machines in a mixed env. this change checks if the DPU that needs updating is in a host with multiple DPUs and will log a debug message.

It also cleans up the extra hosts that the test function that creates a multi-dpu machine leaves around