feat: Add a target field to health probe results
In some cases the same type of health probe checks the
state of multiple components. E.g. a Fan speed probe
might check the state of various fans. Or a disk utilization
probe might check the utilization of different volumes.

So far we used the probe ID field to generate unique
probe IDs for these cases. This however made it harder
to identify failing probes, since a direct string comparision
was no longer sufficient.

The change here introduces an optional `target` field
for probes, which can be used to identify the target
component if the same probe checks the state of multiple
components.