feat: machine-a-tron as a k8s service
This enables support for running machine-a-tron as a deployment within a
kubernetes cluster. Documentation on how to use it is in the machine-a-tron
README.

To make this work:

- Make machine-a-tron able to launch a single BMC mock that delegates to
  backing mocks based on the forwarded IP. The forwarded IP is taken
  from the HTTP `Forwarded: host=<host>` header. Backing mocks are
  registered by MAT's state machine when the BMC gets a DHCP IP.
- Configure carbide's redfish client pool such that, if we're configured
  to override all redfish requests to a single IP, to send `Forwarded:
  host=<ip>` with the original BMC IP.
- Define config options to control this behavior
- Make integration tests use this same behavior (dropping
  MachineATronBackedRedfishClientPool because it's now unused.)
- Build machine-a-tron into a docker image so that it can be deployed
  over k8s

There will be another MR into forged which will define the kubernetes
configs necessary to run this. All existing use cases of machine-a-tron
will continue to work correctly, this is a purely additive change.