fix(api): update IpAllocator to allocate the next available ${prefix_length}
*Note: This is a noop for existing instance and machine interface address
allocations. The `IpAllocator` does now hand out `IpNetwork` values, but those
values are coerced into `IpAddr` values where appropriate.*

For FNN, a tenant instance allocation is changing from a `/32` to a `/30`,
which is actually per DPU (and yes, has implications for multi-DPU). The IPs
themselves are going to be:
- A tenant-instance VRF loopback.
- Reserved.
- DPU IP (acts as the gateway IP for the host).
- Host IP (the IP that gets announced for the tenant instance).

This means we need flexiblility in instance address allocation to [continue to]
allocate a `/32`, and have the flexiblity to allocate larger networks per
tenant instance.

So, with this MR, the `IpAllocator` now has support for allocating the **next
available network prefix** (where this MR also leans on previous work that
prepped the `IpAllocator` (and database) to start storing (and returning)
`IpNetwork` entries.

This is the first iteration of this, and much of this will [eventually be]
changed as we work towards [Carbide IP Prefix Management](https://gitlab-master.nvidia.com/nvmetal/designs/-/merge_requests/25), but for now,
this will allow us to move forward with minimal changes to the codebase,
and, almost more importantly, to the UI, while allowing multi-DPU (and
associated FNN template change) efforts to continue forward.

Some callouts are:
- There will be a subsequent MR to wire this up with the template + agent work
that Bill has, which will include plumbing data via `InstanceNetworkConfig`,
including the requested `prefix_length`, and the resulting prefix allocation.
- We can still do operations by the `address` column of the
`instance_interfaces`, since that is always going to contain the primary
instance address (even in the case of a `/30` allocation). This is a big part
of what helps in terms of backwards compatibility (including DHCP).
- I added a lot of comments to things throughout all of this, including
comments to existing code.

Lots of tests added, some updated, and some continued to work in-place.