refactor: Finish actor model for machine-a-tron
In !2604, I made an actor interface to DpuMachine so that mock host machines
could send messages to their mock DPUs while the control loop was
running. I had the same thing implemented for HostMachine, but it wasn't
needed yet, so I removed it to keep the MR small.

There's a few uses we have for it coming up though: one is a potential
web/API interface to a running machine-a-tron instance, which will need
to be able to do things like trigger a reboot or query information about
HostMachines. Another is that we will need the observed Machine ID of a
running Host Machine from outside its control loop to allow instance
creation without additional bookkeeping.

Some implementation notes:

- HostMachines now construct DpuMachineActors as part of their
  `::new()`, to eliminate the complexity around a HostMachine holding
  either a DpuMachine or a DpuMachineActor based on whether it's started
  or not. To prevent issues where the DpuMachine starts its control loop
  before we start the HostMachine, introduce a `paused` boolean which
  will make `process_state` a no-op until it is unpaused. Do a similar
  thing for HostMachine itself so that callers can `start()`
  HostMachines early (to turn them into HostMachineActors). (A paused
  concept may be useful for a future web UI as well.)

- Move the `stop_when_ready` logic out of HostMachine itself and have
  callers call a simple `.wait_until_ready()` to await readiness (as in
  integration tests). This simplifies some things.

- Make most of the functions in HostMachine/DpuMachine non-public to
  make it more obvious that callers should use the Actor types instead

- Move the force-delete API calls into HostMachine itself, and just call
  them when stopping, which is the logical place to do it (it will also
  make things easier if we want to support adding/deleting machines from
  a future web UI.)