feat: Show aggregate Host health when using GetMachine API
With this change, the GetMachine API behaves as follows:
- When the function is called on a Host Machine, it will return the aggregate health (includes Host Health, DPU health and all overrides) in the health field
- When the function is called on a DPU Machine, it will return only the respective DPUs health - as reported by forge-dpu-agent

This means that forge web UI would now correctly display the aggregate host health.

The change is implemented by deriving the RPC Machine state based on the ManagedHostStateSnapshot - and not just based on the Machine snapshot/data.

The change is not yet implemented for the FindMachines/FindMachinesById APIs - these require further changes.

Note: Downside of this change is that we load DPU related data even if the query is just for a Host - and the other way around. However due to recent efficiency improvements, this should not be a huge issue.