feat: Allow to query Machine history with FindMachinesByIds
With this change, calls of the FindMachinesById API can optionally request the history for each queried machine.

The change is backward compatible, since the field layout of the new query is the same as MachineIdList.

While `include_associated_machine_id` could also be specified in the same request, I opted to keep it at being always `true` to keep the functionality consistent with how the method behaved before.

Loading the information is free for Hosts anyway - and most queries just target hosts.