test: Add instance creation to machine-a-tron integration test
This cherry-picks some changes out of !2680 to make that MR smaller.

- Support waiting on any state, not just Ready
- When a host reaches Ready, allocate an instance (only in the
  integration test code path... for the TUI path, callers can press "i"
  to allocate an instance)
- support callers waiting for all hosts to be ready, without stopping
  the machines yet
- Have the calling test code make assertions that the instance is
  allocated and looks as expected, before stopping/deleting the hosts

Also, clean up some unnecessary complexity in MachineStateMachine...
since we now support the state machine returning "sleep for this
duration", there's no reason to keep track of the last time we ran the
network status update (and thus no reason to support a
change-state-and-sleep state). The MachineUp state is now a lot simpler.