feat: FORGE-3912 Include gateway in InstanceInterfaceStatus
This plumbs the gateway field from InstanceInterfaceStatusObservation all the
way to the InstanceInterfaceStatus RPC object so that the tenant UI can see it.
The bluefield agent will populate this from the interface config it gets from
ManagedHostNetworkConfig.

We could think about falling back on fetching the value from the
network_prefixes table if the agent is too old to have sent it, but there are a
lot of baked in assumptions that a InstanceInterfaceStatusObservation can be
trivially converted to an InstanceInterfaceStatus, and implicitly fetching from
the database to do the conversion seems like it's more trouble than it's worth.