fix: refine gpu pci id check in hw enumeration
The gpu pci check was originally added to work around nvidia-smi
failing if no GPU was present, as the nvidia module was not
loaded by the kernel. This commit further refines the check such
that it finds any nvidia owned GPUs that are loaded by the nvidia
kernel module and does not rely on PCI ID values.

Ref: https://nvbugspro.nvidia.com/bug/4756066