fix(api): measured boot cleanup -> CarbideError, CarbideResult, etc
Still banging away at some post-commit cleanup. I had dropped a lot of
use of `eyre::eyre!` everywhere, and I'm in the process of cleaning it
up by correctly using:
- `DatabaseError`
- `CarbideError`
- `CarbideResult`
- `CarbideCliError`
- `CarbideCliResult`

I basically just pick a starting point, and follow where it takes
me. There are still more to do here, but this is a checkpoint so
I can go back to some FNN [for multi-DPU] work.