fix(api): preparing FNN prefix allocation support for tenant instances
Historically, tenant instances have been allocated a single `/32` IP address
per {physical, virtual} function on their DPU.

With [FNN](https://gitlab-master.nvidia.com/nvmetal/designs/-/blob/main/designs/0011-fnn-architecture.md),
that is changing, where tenant instances will now be allocated a network
*prefix* per PF/VF on their DPU. And, in the case of multi-DPU, on their *DPUs*.

And this is the preface to this change. We need to make changes to the
`IpAllocator` to start allocating *prefixes* (of possibly varying lengths,
but `/30` to start) instead of just a `/32`.

In preparation for that, this means we need to introduce `IpNetwork` in places
where `IpAddr` is used.

Thankfully, both can co-exist, and not only does the *Postgres* `inet` type
support prefix lengths (and, as such, can bind to an `IpNetwork`), but both
`inet`/`cidr` and `IpAddr`/`IpNetwork` all play very nicely together; `inet`
is basically an `IpNetwork` with an implied default of `/32`, but WITHOUT
netmask enforcement (`cidr` is actually better in this case).

So, this alters the `instance_addresses` table to add a `prefix` column,
whose value == the `address` column, but with an explicit prefix length
set, and will subsequently begin populating that column going forward.

To start, it will simply be a `/32`, but a subsequent MR which introduces
an `IpAllocator` change will allow that to become configurable, at which
point the `address` column will become the PF/VF address allocated to
the DPU, and the `prefix` column will be the entire prefix (whose other
IPs will be used for other components of FNN for the DPU).

The intent is to make this a pretty close noop/drop-in to start, with the
follow-up `IpAllocator` MR to take things further.