feat: Support releasing instances in machine-a-tron
Currently machine-a-tron will get stuck in the HostCleanup state when
trying to deallocate an instance.  The easy fix is to simply send a
`cleanup_machine_completed` action when this happens, but currently the
state machine does a lot of guess-work as to when to perform various
actions... there's very little paying attention to the ForgeAgentControl
action the server wants us to do.

So, overhaul the MAT state machine to more closely resemble a real
machine's lifecycle, keeping our business logic out of the state
machine, and instead keep track of what image the PXE server is telling
us to run, what image is installed, and try our best to emulate the
behavior of that image.

Details:

- Add a booted_os field to the MachineUp state so that we know what
  behavior to emulate. Move all the hardware discovery stuff to
  MachineUp, since we can do that as part of the OS we're emulating
  instead of hardcoding it into the state machine

- Parse the PXE URL and determine whether we should boot to carbide.efi
  (forge agent) or scout.efi (scout), versus the installed OS (exit)

- Keep track of what OS is installed, so that if PXE returns "exit", we
  can boot to that. For DPUs it means running the forge agent service
  when PXE returns exit (but only if we've run the carbide.efi image
  once, which is supposed to install it), but for hosts it means running
  nothing. Return errors if a DPU tries to boot to an empty OS, or if a
  Host tries to boot to the DPU OS.

- When emulating scout, poll for the next ForgeAgentAction and perform
  it (this is where the fix comes in, because the action may be running
  HostCleanup.) Move the MachineValidation logic to this too.

- Integration tests: Make it so machine_a_tron::run_local() doesn't
  add/remove an instance, but instead let the integration test itself do
  that as part of the other assertions it's doing (and let the test
  determine when to stop the mocks.) This means we use grpcurl to
  allocate/release an instance the same way the other integration tests
  work, and we no longer have to use MAT's
  api_client::allocate_instance.