feat: pre-merge test
MVP

Adds a new GitLab CI job pre-merge-test to Merge Request pipelines where
specified directories are changed.

The job runs on a specific gitlab runner machine provisioned
specifically for this purpose with the "dev env" image from
https://gitlab-master.nvidia.com/nvmetal/ci-tools/internal-services-images.
(This is similar to the pre-merge test in `forged` as added here: nvmetal/forged!1896)

The job provisions the dev environment k3s cluster from `forged` and
then builds the carbide images and deploys them to the cluster.

We then run two validations:

- the existence of the carbide pods we expect to see
- the health status of all pods, deployments, services, etc. as reported
  by kubernetes.

Future enhancement: connect this dev environment to an Acorn host and
run the existing machine-lifecycle test.