fix(fmds): Update fmds urls to conform with ec2 service
The ec2 cloud init code is very specific in what URLs it expects if we
want cloud init to autodetect. This commit moves the user-data out of
meta-data and puts it in its own top level router for cloud init. It
also adds a meta-data route, which will cause cloud init to fail if it
is not present. It also adds a route for a bug in cloud-init,
meta-data/ with a trailing slash. The trailing slash caused a move of
how the urls were constructed, which caused the large refactor of the
code.