chore(api): drop DbMachineId in lieu of using MachineId
I think there might have been some ideology here, which I'm totally
in support of, to have a `DbMachineId` which lives in `db/`, whose
purpose was to be able to wrap + bind a `MachineId` (which lives in
`model/`. We didn't want to cross streams between `db` <-> `model`
like that (or import `sqlx` into `model`), so being able to build
a `DbMachineId` in `db/` and then derive a `MachineId` model from
it totally makes sense.

...but, I ALSO think, if the path forward is to move ID-types of
things out into its own `common/forgeid/` (or some type of similar
crate), that it's reasonable to fold the `sqlx`-binding type things
into `MachineId`, and get rid of a number of extra operations (and
extra `FromRow` implementations) by just having a singular `MachineId`
that does it all.

..so, I'm presenting this MR as moving toward the path to that. I'm
happy to have it smacked down (or at least deferred until the wider
effort to move IDs to `common/` happens), but this would eventually
happen at that point anyway!

And this is my "I want to write an MR tonight" of the night.

Enjoy!