chore(api): give ConfigVersion some sqlx bindings, integrate accordingly
My goal is/was to get rid of some low-hanging `FromRow` implementations
for places that just need some simple bindings, and I noticed `ConfigVersion`
was a good candidate to start with.

This adds `sqlx` bindings to `ConfigVersion`, so we can just insert/fetch
database rows with a `ConfigVersion`-backed column, and have it do what we'd
expect (getting rid of a bunch of `"*_version_str"` examples in the
process).