fix: Machine-a-tron: Don't return error if DPU boots without PXE
If a mock DPU boots without carbide booting it to PXE yet, we were
throwing an error... but I think this is possible to happen in some
cases (ie. carbide seeing a DPU and not necessarily booting to
carbide.efi the first time it boots). And regardless, a real DPU
wouldn't just continually boot in a loop in this case, we would likely
just not hear from it until we rebooted it from the BMC, which
machine-a-tron should emulate by doing nothing.