fix(api): introduce VpcVirtualizationType::{Classic,FnnL3} to the fray
Continuing on moving on towards getting FNN (and corresponding
NVUE template changes) integrated with everything. Based on feedback
from Bill (and seeing how the `forge-dpu-agent` main loop does a
`match` based on `NetworkVirtualizationType` to decide how to `update_nvue),
it makes sense that this would exist for the purpose of that (and honestly,
it makes sense that we'd have this anyway).

This doesn't actually get used anywhere yet (and I have the agent `panic!`
if it sees something -- this will get filled in as part of https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/2502 updates).

Dropping a `TODO(chet)` to also fold `VpcVirtualizationType` and
`NetworkVirtualizationType` into a `common/network` crate (or something
similar) -- I think the only reason they're separate right now is so
we don't introduce a dependency from `bluefield` -> `api`.