fix(build): make sure to `docker run --pull always` for `:latest` tags
We ran into an issue where people were trying to do builds of various
things, and it was using a `:latest` image that didn't have the latest
changes in it.

It turns out that it was because their environments had pulled `:latest`
a long time ago, so `docker run` was just using that (and not actually
pulling down `:latest`).

This changes the tasks to `docker run --pull always` for `:latest` cases,
ensuring the `:latest` is actually the latest `:latest`.

This also explicitly sets the `:latest` tag for instances where no tag
is set (and `:latest` is implied).