fix(api): sync timestamps in `ReprovisionRequest`
Noticed this calls `chrono::Utc::now()` separately within the
same message. It's probably close enough that it's probably
the same (I'm not sure of the clock resolution tbh), but this
will ensure it's the same time for both, which is probably
what makes sense here (I think it could get confusing otherwise).