feat: Support NetworkDeviceFunctions in mock_machine_router
New changes landing in libredfish need information from the
NetworkDeviceFunctions URIs for each DPU. Since we build mock data
manually for DPU information, we can't just fall back on the upstream
tar file, we have to be able to rewrite the response with our own
MAC/Serial/etc information.

So this makes it so we fetch a tar entry which is known to have good DPU
mock data (for NIC.Slot.5 within the dell mock) and rewrite portions of
the data for what we're mocking.

This combined with !2762 should make the new libredfish code work in
machine-a-tron, which will unblock !2752.