feat: Support $expand=.($levels=1) in bmc mock
This makes bmc-mock use a new wrapper router, which supports expanding
requests that pass `?$expand=.($levels=N)` in the query string, which
according to the redfish spec is supposed to cause "Members" links
(possibly others, but we're only supporting Members for now) to expand
to the full objects in a single request, instead of containing just
their links.

The implmementation is compeltely generic and can wrap any axum router.
It will only perform replacement if there is JSON in the wrapped
request, and if that JSON has a "Members" section that has OID's, etc.