fix(api): plumb through the allocated tenant instance prefix to the DPU
Currently, in the `forge-dpu-agent`, we're statically deriving the allocated
`network` based on the allocated instance IP address (e.g. we're straight up
doing string concatenation of `ip + "/32"`).

This changes things so now we pass down the instance-allocated `prefix` stored
on the Carbide side (which is also known as the `prefix` column in the
`instance_addresses` table) -- for non-FNN interfaces, the `prefix` is *exactly*
as we hard-code it today with string concatentation (`<ip>/32`), meaning this
change is a drop-in for (AND we can drop the concatentation).

Then, when FNN-specific instance allocations start happening, we'll be able
to drop in the `instance_prefix` where it goes.

This includes changes to `update_nvue` to populate the existing template
with the `instance_prefix` value we [now] get from the `FlatInterfaceConfig`
(which is derived from an `InstanceInterfaceConfig`), as well as updating
`InstanceInterfaceStatus` and `InstanceInterfaceStatusObservation` to feed
the config back to *Carbide*.

So, this in itself is a noop, but is building the path towards the upcoming
FNN {Classic, L3} template populatation (per https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/2502).

And actually, I think this is the last MR we'll need before it's time to glue
things together with that aforementioned MR.