chore(api): tweaks to db delete query execution
I went to go double check that there weren't any other cases of
deletion failures being silently dropped, and while I was in here,
realized I kind of like doing a `.map(|_| ())` so we can just return
the result from the `sqlx::query_as` directly (without needing to `?`
or end with an `Ok(())`, so I'm proposing that. :P It seems to look
cleaner, and also works nicely with getting rid of the dropped `let _xxx`
assignments.