fix(api): introduce default values when `interface_prefixes` aren't set yet
For existing instances, the `InstanceInterfaceConfig` (backed by the
`network_config` column in the database), is not going to have any
`interface_prefixes` set yet (since this will only happen for newly
allocated instances which go through `Instance::update_network_config`).

This updates `InstanceInterfaceConfig` to allow for a default value
of an empty set of interface prefixes (for deserialization from the
database).

Additionally, this also updates the `FlatInterfaceConfig` population
for both `admin_network` and `tenant_network` calls (which happens
when the agent does a `get_managed_host_network_config`) to check for
a missing `interface_prefix`, which will now infer a default value
of `/32`.

I *did* leave the `validate()` check in place with a `TODO`, even
though it's kind of silly now. Basically, if there happen to be
interface prefixes configured for interface, it means it's a new
instance that is leveraging the new flow, so we can actually check
to validate that the network prefix ID keys match across both.