feat(dpu-agent): Add metrics for machine and agent start time
We currently don't have a good way to tell if a DPU was rebooted
(unexpectedly or unexpectedly), or if the agent process is crashing
repeatedly.

This adds two new metrics, `machine_boot_time_seconds` and
`agent_start_time_seconds`, both of which are Unix timestamps and are
recorded once at agent startup time.