fix(admin-cli,api): support for updating the VPC virtualization type
This updates the `UpdateVpc` API call (and the `forge-admin-cli`) to
allow for updating the current VPC/network virtualization type. The
primary driver for this is so we can configure a new VPC as `fnn-l3`
or `fnn-classic` without needing UI support to do so (and not needing
to manually go into the database to do so), but is also something
that makes sense to allow an operator to change if needed.