fix(api): add support for the GNI-supplied internet L3VNI
This adds support for setting a side-wide L3VNI, which is supplied
to us by GNI, for allowing VPCs to access the Internet. If unset
for a given site (since this is optional to configure), VPCs in that
site will be unable to access the Internet.

I did drop a `TODO` to consider making this configurable per-VPC,
as in:
- Being able to set a site-wide value, but being able to flip it off for a given VPC.
- Or have the site-wide value, and require a tenant to select yes/no for Internet.
- Or having no site-wide value set, and allow per-VPC enablements with custom values.