fix(dpu-agent): wiring VPC/L3VNI in with the nvue templating configs
This wires up the L3VNI value that comes from the API into the template
input configs.

What's nice was, as I started digging around, I saw there was already template
support for it, and a way to inject it via the CLI options using `ct_l3vni`,
so I just had to connect in the value from the API (thanks Graham)!

..and this is when I also saw that we're setting the `L3VNI` at the
`ComputeTENANT` level in the configs for now (and not in the `PortConfigs`),
so this updates `ManagedHostNetworkConfigResponse` to include the `vpn_vni`
(the same one in the `FlatInterfaceConfig`) so we can leverage it in at
both the `ComputeTENANT` and `PortConfig` levels.