feat: Caller can choose boot interface
Upgrading libredfish to 0.23.1. `forge_setup` now takes an optional MAC
address. If given it will set that as the first boot device. If not
given (previously the only option) it looks for a Mellanox Bluefield DPU
and uses that to boot from.

This is to support multi-dpu and zero-dpu.

That paragraph above is why this commit exists. The bulk of the code in
here isn't related to that at all.

The Rust async ecosystem recently underwent a big change when `hyper`
upgraded from 0.14 to 1.0. It is very backwards incompatibler.

`axum` (0.6 -> 0.7), `reqwest` (0.11 -> 0.12), and `http` (0.2 -> 1.0)
because they depend on `hyper`, also have backwards incompatible versions.

carbide-api is still on the old side of the river but libredfish has
now crossed. That takes a bit briding. The biggest change is we now
don't use `http` directly, except when talking to libredfish. We
always use the vendored version in hyper, `hyper::http` because
that's the old and still compatible one.

Separately, current `sha2` is not compatible with the upgraded
`reqwest` (from libredfish), so we use a pre-release sha2.