feat: Add a target to HeartbeatTimeout alerts
Originally HeartbeatTimeout only signaled that a DPU was not responsive.
After we added hardware health reports, a HeartbeatTimeout could however also signal that the hardware-health service is not responsive.

In order to allow to distinguish what is wrong,
we now indicate the component (`hardware-health` vs `forge-dpu-agent`) inside the `target`field of the alert.