fix(api): implement From trait for MachineId -> rpc::common::MachineId
I noticed `LinkedExpectedMachine` was using an `Option<String>` for its
`machine_id` (instead of a `MachineId`), so I changed it to be an
`Option<MachineId>`, but then I noticed how it was building an
`rpc::common::MachineId` (which was odd), which was because I hadn't
implemented `From` yet for it.

This does that, and subsequently updates a number of other places
which benefit from the new trait implementation.