fix: Fix DPU NicMode detection
We were looking in Bios.NicMode, but the attribute is actually
Bios.Attributes.NicMode, you can see with:

```
tar -xzOf dev/bmc-mock/nvidia_dpu.tar.gz \
redfish/v1/Systems/Bluefield/Bios/index.json | jq .
```

Also add machine-a-tron support for putting DPUs in NIC mode, and make
an integration test to assert on it. The assumption is that the BMC
information for DPUs in NIC mode will be identical except for the
Bios.Attributes.NicMode setting will be set to "NicMode" instead of
"DpuMode". With a single DPU in NicMode, this asserts that it behaves
the same as a zero-DPU host.

Unrelated: Change each machine-a-tron integration test to use a single
host, as testing with 10 hosts each is slower, needlessly chatty in the
logs, and doesn't really buy us much.