fix(carbide-web): Show proper DPU health status on DPU Network Status page
The previous change https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/2819 transitioned the DPU Network Status page to show the new health report format, instead of the legacy format which will be phased out.

This change along was not enough: Since the DPU health is stored in a different place in the database, it was not properly returned to the API.

This change makes the status page load the full DPU information via FindMachines, which has the necessary information.

In future changes we might want to phase out reading back the `DpuNetworkStatus` type in the web-ui for using a different API or types which are read-optimized (like FindMachines). `DpuNetworkStatus` was mostly intended as a way for the DPU to submit its report.

The change also includes displaying the version of `forge-dpu-agent` that is stored in the Machine Inventory and is therefore consistent with other APIs. We can work towards no longer persisting the agent version in `model::network::MachineNetworkStatusObservation`.