Chore: Remove non existent error condition in discovery
In doing some research to how discovery works these days, it became
apparent that the first step of the old two step process, run, was no
longer used. For some reason it still existed but was just always
returning OK. This led to a bit of code that bubbles an error up to
the API no longer ever being exercised other than in tests. This
commit cleans up the no longer existent error condition so that it
just does not exist in the API anymore.