fix: Coalesce and throttle machine-a-tron get_machine calls
Instead of having each machine call the API on its own to get the API
state, coalesce the calls into a single call to GetMachines with a list
of machine_ids.

Also, make sure DpuMachines only refresh their state on an interval as
well... their state was updating every call to process_state.