fix(agent,api): configure + pass `svi_ip` down into the nvue `PortConfig`
There has been an empty `svi_ip: None` placeholder in `update_nvue` for a
bit now (with template support for an `.SviIP`, while we waited for things
to eventually start to get some eyes on it -- and here we are!

This will parse the `svi_ip` (aka the gateway IP) out of the allocated
`interface_prefix` and pass it through via the `FlatInterfaceConfig`,
where it then makes its way into an `nvue::PortConfig`, where it then
makes its way into the `TmplConfigPort`.

A new function called `get_svi_ip`, which lives alongside `get_host_ip`,
is what is being used to parse the IP out from the interface prefix.

This continues to stay in-line with doing as much as we can server-side,
and then passing baked config down to the agent.