chore: Stop persisting DPU health in `MachineNetworkStatusObservation`
The DPU health is already stored in `Machine::dpu_agent_health_report` in the new format.
There was only two references in the update workflow and DPU metrics to the old format. This has been changed as part of this merge request.

Therefore we can now stop persisting the DPU health in this field.