feat: Emit probe ID and target as aggregate Health and DPU health attributes and fix most multi-dpu metrics
With this change, every host and dpu health alert metric will carry a `probe_id` and `target` field which makes it easier to identify the problem just based on the metric.

For the DPU health probe alerts, the change will
additionally retain the old `failure` attribute - which will now contain both the probe ID and target (if existing).

In addition the change will correctly emit most per-DPU metrics (e.g. dpu health alerts and software versions) in case of multi-dpu systems.

The only metric which is not correctly emitted for multiple DPUs is the client certificate lifetime metric.