chore(ci): remove conventional commit enforcement
This removes conventional commit ennforcement and we're, instead,
curating a CHANGELOG.md file and expecting MR authors to update the
change log.  Therefore there's not a lot of value to conventional
commits as we won't be auto-generating changelogs.

This does mean people need to write effectively their changelog either
in the git commit message or in the CHANGELOG.md.  But that had to
basically be done anyway.