chore: remove vpc/ from .gitignore
This is from a time when there was a subsystem called `vpc`, which was part of the original DPU agent, which was written in Golang, and the agent + Carbide communicated via k8s. It used schema files and code generation, and the purpose of this was to hide those generated files.

I was recently trying to add a directory named `vpc` nested a few levels deep, and `.gitignore` was catching it. Rather than tweak to `./vpc`, I'm just removing it since it's no longer needed.

Thanks to @meinwag and @grahamk for the history lesson on that one!

See merge request nvmetal/carbide!2886