pre-merge test: merge trunk before running test
Attempt to bridge the coverage gap cause by the carbide MR pipeline not requiring the branch to be up-to-date.

The pre-merge test will now merge trunk in before executing the test. 

If there's a merge conflict, the job will exit. The developer must update their branch and resolve the conflict before the job will run.

See merge request nvmetal/carbide!2888