fix: only check if the site-wide bmc root credentials are valid in site-explorer.
fix: only check if the site-wide bmc root credentials are valid in site-explorer (checking if the username and password in the Vault entry are not empty strings). The other site-wide credentials contain a username entry that is never set (because the uefi menu doesnt prompt for a username).

See merge request nvmetal/carbide!2887