fix(api): move MachineId to the new common UUID location
_This is a follow up to https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/2884, which moved over some initial IDs into the new location._

So this turned into quite a treat -- first, of course, `MachineId` is used all over the place, so there was that. And then there were:

* Some trait implementations that `MachineId` was using (one was for displaying tables in the CLI, c/o some measured boot stuff I did).
* Some shared stuff that was related to the admin CLI, which I had to move into `common/utils/admin_cli`.

So those all got updated in the process as well.

In the end, `MachineId` is now in `common/uuid` with a bunch of other IDs that have already been moved over, and the measured boot IDs are next.

See merge request nvmetal/carbide!2901