chore: Remove `Machine::associated_host_machine_id` and `associated_dpu_machine_ids` fields
The fields are not stored directly in the database. They are a derived from
the MachineInterface relations.

This change reflects this, by making `associated_dpu_machine_ids` a function
which retrieves the IDs from the `interfaces` field.
It avoids having inconsistent data stored in `interfaces` the the fields.

See merge request nvmetal/carbide!2905