fix(api): save some vpc db calls when building FlatInterfaceConfigs
https://jirasw.nvidia.com/browse/FORGE-4348

This came out of the FNN template MR -- the only caller of `tenant_network`, which builds the `FlatInterfaceConfig` for a given interface, already has the VPC the allocated instance is in, so lets not make a database call to get the VPC information again.

Instead, just pass the `vpc_vni` and computed `vpc_prefixes` into it, and let `tenant_network` deal with the rest of the \[non-VPC related\] things.

This also drops some `mut` things, which is always nice.

See merge request nvmetal/carbide!2914