feat: Transform StateHandlerOutcome into an error if object stays in state for...
feat: Transform StateHandlerOutcome into an error if object stays in state for longer than the defined SLA

The state handler can return one of these outcomes:
- Ok(`Transition`) to transition between states
- Ok(`Wait`) to wait for a certain system condition to become true
- Ok(`DoNothing`) to indicate waiting for a user interaction
- Err(e) to indicate an error during state handling

When the state handler returns `Wait` or `DoNothing`, the object simply stays in its current state.
This is usually ok for some time. But after a while the object might be "stuck". It should have
transitioned while waiting for a condition, but it did actually not.

We already have a metric for objects that are in state for longer than the SLA,
but the property did not yet show up anywhere else.

This change transforms StateHandlerOutcomes that suggest staying in the current state
into StateHandlerOutcomes that describe an error in the case the SLA has elapsed.
This will make stuck objects visible in 2 additional ways:
- the state handler error metric ( e.g. `forge_machines_with_state_handling_errors_per_state`) will show Machines with an error of `time_in_state_above_sla`
- The State Handler Outcome details in the API and Web UI will show `The object is in the state for longer than defined by the SLA. Handler outcome: {handler_outcome}"`

See merge request nvmetal/carbide!2913